Add detailed monitoring pages for CPU, Disk, Memory, and Network
- Implemented CPU details page with real-time usage chart and per-core statistics.
- Created Disk details page featuring usage metrics, pie chart for space distribution, and partition information.
- Developed Memory details page displaying current usage, breakdown chart, and top memory-consuming processes.
- Added Network details page with throughput chart, upload/download rates, and interface statistics.
- Integrated real-time data fetching and chart updates for all monitoring pages.
- Enhanced dashboard with links to detailed views and improved layout for better user experience.

+++ internal/models/process.go
@@ -3,7 +3,7 @@ package models
 type ProcessStatus struct {
 	PID        int32   `json:"pid"`
 	Name       string  `json:"name"`
-	CPUPercent float64 `json:"cpu_percent"`
+	CPUPercent float32 `json:"cpu_percent"`
 	MemPercent float32 `json:"mem_percent"`
 	Status     string  `json:"status"`
 }
